requirements.txt
altgraph==0.17
appdirs==1.4.3
beautifulsoup4==4.8.2
bottle==0.12.18
bottle-websocket==0.2.9
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cssselect==1.1.0
Eel==0.11.0
fake-useragent==0.1.11
future==0.18.2
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.8
lxml==4.5.0
parse==1.14.0
patool==1.12
pefile==2019.4.18
pycparser==2.19
pyee==7.0.1
pyppeteer==0.0.25
pyquery==1.4.1
pywin32-ctypes==0.2.0
requests==2.22.0
requests-html==0.10.0
selenium==3.141.0
six==1.14.0
soupsieve==1.9.5
tqdm==4.42.1
urllib3==1.25.8
w3lib==1.21.0
webbot==0.1.4
websockets==8.1
whichcraft==0.6.1
